TST Add test file test_messages
Currently contains one test case for testing the
willChange and deprecated decorator.
